+++ src/utils/formatError.js
@@ -1,10 +1,10 @@
 export default class {
-  constructor ({ errors, status = {} } = {}) {
+  constructor ({ errors, status = { code: 400 } } = {}) {
     return {
       response: {
         data: {
           errors,
-          status: { code: 400 }
+          status
         }
       }
     }