+++ src/vuexOffline.js
@@ -34,11 +34,9 @@ export default class VuexOffline {
     const collectionHandler = new CollectionHandler(collection)
     const { filters: filtersList, search: searchList } = collectionHandler.getFiltersAndSearch()
     const fieldsList = collectionHandler.getFiltersFields()
-    const fieldsWithRelation = collectionHandler.getFieldsWithRelation()
     const allFields = collectionHandler.getAllFields()
 
     const relationsHandler = new RelationsHandler(collection, this.databaseSetup.collections)
-    const fieldsWithRelationOptions = await relationsHandler.getFieldsWithRelationOptions()
 
     const nested = new Nested()
 
@@ -133,7 +131,7 @@ export default class VuexOffline {
 
         replaceItem (state, payload) {
           const index = state.list.findIndex(item => item[idAttribute] === payload[idAttribute])
-          
+
           ~index ? state.list.splice(index, 1, payload) : state.list.push(payload)
         },
 
@@ -194,6 +192,8 @@ export default class VuexOffline {
         },
 
         fetchSingle: async ({ commit }, { form, id, params, url } = {}) => {
+          const fieldsWithRelationOptions = await relationsHandler.getFieldsWithRelationOptions()
+
           if (!id && form) {
             return {
               data: {
@@ -215,7 +215,7 @@ export default class VuexOffline {
             const parsedDocument = document.toJSON()
             const fields = form
               ? fieldsWithRelationOptions
-              : await relationsHandler.getFieldsWithRelationOptionsById(parsedDocument)
+              : await relationsHandler.getFieldsWithRelationOptionsById(document)
 
             commit('replaceItem', parsedDocument)
             commit('setErrors', { model: 'onFetchSingle' })
@@ -252,6 +252,7 @@ export default class VuexOffline {
           { filters = {}, increment, ordering = [], page = 1, limit, search } = {}
         ) => {
           try {
+            const fieldsWithRelationOptions = await relationsHandler.getFieldsWithRelationOptions()
             const filtersHandler = new FiltersHandler({
               receivedFilters: filters,
               filtersList,